+++ exosphere/build-tools/script-tests/fixtures/pass-angular/src/app/app.module.ts
@@ -1,0 +1,17 @@
+import { BrowserModule } from "@angular/platform-browser";
+import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
+
+import "@boomi/exosphere/dist/styles.css";
+import "@boomi/exosphere/dist/icon.js";
+import "@boomi/exosphere";
+
+import { AppComponent } from "./app.component";
+
+@NgModule({
+  declarations: [AppComponent],
+  imports: [BrowserModule],
+  providers: [],
+  bootstrap: [AppComponent],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
+})
+export class AppModule {}