+++ exosphere/build-tools/script-tests/fixtures/pass-react/src/main.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+
+import "@boomi/exosphere/dist/styles.css";
+import "@boomi/exosphere/dist/icon.js";
+
+import App from "./App";
+
+ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);